Attempt to address CI golden files spec failures
The failure is not reproducible in my local environment. It appears
to happen because we have multiple of these type of calls in the
same method:

```ruby
expect(FirewallRule).to receive(:generate_uuid)
```

And it isn't using the latest definition. The CI failure is a
Sequel::UniqueConstraintViolation using a uuid specified in
the above method call on line 19, when there is a later method
call on line 37 that should override that.

This seems like a bug in rspec. Try working around it by using
a single call with more arguments.

That fixed one issue, which occurred during test setup, before
running the tests on the golden files.  The next issue was the
reuse of an id during the file testing.  Attempt to fix that by
adding 6 more FirewallRule ids to the rotation.